Create playRound function
Create the playRound() function that determines the outcome of the game
based on the playerSelection and computerSelection arguments.

The function utilises if/else conditional statements and returns a
string in the console that states who wins the round.

Beforehand, a prompt for user input was created, and a function to
capitalise that string is used to make the value remain consistent with
the playRound() conditionals.