Change player input from prompts to buttons
The player input was changed from prompts to buttons, which call
playRound() based on the textContent of the button that was pressed.

The loop in game() that plays 5 rounds was removed. The game() and
promptPlayer() functions are left alone for now, in case the code can be
used for other functionalities.