Display outcomes on webpage instead of console
The coding was reworked so that the outcomes of each round are displayed
on div elements rather than in the console, where it previously showed
the results. One of the elements displays the outcome of a round, and
the player and computer scores are updated after every round. The scores
reset after either the player or computer reaches 5 points.

The game() and promptPlayer() functions were deleted, as they were not
being used anymore.