Make CSS improvements to the results sections
Added a simple animation to the results section so that it jumps
every time the player presses a button. The animation should restart
every time a button is pressed, so resetAnimation() was created to fix
this.

Additionally, the background colours of the score sections were changed
to distinguish them from the buttons. The HTML for these sections were
also rearranged to accommodate future CSS changes.